All methods separate to 2 class: DuckClient and DuckActionsClient.
