+++ src/test/java/autotests/clients/DuckActionsClient.java
@@ -18,54 +18,6 @@ public class DuckActionsClient extends TestNGCitrusSpringSupport {
     @Autowired
     protected HttpClient yellowDuckService;
 
-    @Description("Эндпоинт, создающий уточку")
-    public void createDuck(TestCaseRunner runner, String color, String height, String material, String sound, String wingsState) {
-        runner.$(http()
-                .client(yellowDuckService)
-                .send()
-                .post("/api/duck/create")
-                .message()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body("{\n" +
-                        "  \"color\": " + "\"" + color + "\",\n" +
-                        "  \"height\": " + height + ",\n" +
-                        "  \"material\": " + "\"" + material + "\",\n" +
-                        "  \"sound\": " + "\"" + sound + "\",\n" +
-                        "  \"wingsState\": " + "\"" + wingsState + "\"\n" +
-                        "}"));
-    }
-
-    @Description("Эндпоинт, удаляющий уточку")
-    public void deleteDuck(TestCaseRunner runner, String id) {
-        runner.$(http()
-                .client(yellowDuckService)
-                .send()
-                .delete("/api/duck/delete")
-                .queryParam("id", id));
-    }
-
-    @Description("Эндпоинт для получения списка id всех уточек")
-    public void listDuckId(TestCaseRunner runner) {
-        runner.$(http()
-                .client(yellowDuckService)
-                .send()
-                .get("/api/duck/getAllIds"));
-    }
-
-    @Description("Эндпоинт для обновления характеристик уточки")
-    public void updateDuck(TestCaseRunner runner, String color, String height, String id, String material, String sound, String wingsState) {
-        runner.$(http()
-                .client(yellowDuckService)
-                .send()
-                .put("/api/duck/update")
-                .queryParam("color", color)
-                .queryParam("height", height)
-                .queryParam("id", id)
-                .queryParam("material", material)
-                .queryParam("sound", sound)
-                .queryParam("wingsState", wingsState)
-        );
-    }
     @Description("Эндпоинт, заставляющий уточку лететь")
     public void duckFly(TestCaseRunner runner, String id) {
         runner.$(http()
@@ -94,7 +46,6 @@ public class DuckActionsClient extends TestNGCitrusSpringSupport {
                 .queryParam("repetitionCount", repetitionCount)
                 .queryParam("soundCount", soundCount));
     }
-
 
     @Description("Эндпоинт, заставляющий уточку плыть")
     public void duckSwim(TestCaseRunner runner, String id) {
